+++ R/utils.R
@@ -1089,8 +1089,11 @@ UpdatePrepList <- function(PrepList, pars, random, lr.random, lrPars = list(), M
     pars$value[pars$name %in% c('g', 'u')] <- logit(pars$value[pars$name %in% c('g', 'u')])
     pars$lbound[pars$name %in% c('g', 'u')] <- logit(pars$lbound[pars$name %in% c('g', 'u')])
     pars$ubound[pars$name %in% c('g', 'u')] <- logit(pars$ubound[pars$name %in% c('g', 'u')])
-    if(PrepList[[1L]]$nfact > 1L)
-        PrepList[[1L]]$exploratory <- all(pars$est[pars$name %in% paste0('a', seq_len(PrepList[[1L]]$nfact))])
+    if(PrepList[[1L]]$nfact > 1L){
+        mat <- matrix(pars$est[pars$name %in% paste0('a', seq_len(PrepList[[1L]]$nfact))],
+                      nrow = length(PrepList[[1L]]$K), ncol = PrepList[[1L]]$nfact, byrow=TRUE)
+        PrepList[[1L]]$exploratory <- all(sort(colSums(!mat)) == seq(0L, PrepList[[1L]]$nfact - 1L))
+    }
     ind <- 1L
     for(g in seq_len(length(PrepList))){
         for(i in seq_len(length(PrepList[[g]]$pars))){
@@ -2134,9 +2137,9 @@ computeNullModel <- function(data, itemtype, key, group=NULL){
     itemtype[itemtype == 'Rasch'] <- 'gpcm'
     if(!is.null(group)){
         null.mod <- multipleGroup(data, 1L, itemtype=itemtype, group=group, verbose=FALSE,
-                                  key=key, technical=list(NULL.MODEL=TRUE))
+                                  key=key, quadpts=3, technical=list(NULL.MODEL=TRUE))
     } else {
-        null.mod <- mirt(data, 1L, itemtype=itemtype, verbose=FALSE, key=key,
+        null.mod <- mirt(data, 1L, itemtype=itemtype, verbose=FALSE, key=key, quadpts=3,
                          technical=list(NULL.MODEL=TRUE))
     }
     null.mod
@@ -2239,9 +2242,9 @@ tli <- function(X2, X2.null, df, df.null)
 
 
 rmsea <- function(X2, df, N){
-    ret <- ifelse((X2 - df) > 0,
-                  sqrt(X2 - df) / sqrt(df * (N-1)), 0)
-    ret <- ifelse(is.na(ret), NaN, ret)
+    ret <- suppressWarnings(ifelse( (X2/df - 1) > 0 & df > 0,
+                   sqrt((X2/df - 1) / (N-1)), 0))
+    ret <- ifelse(df <= 0, NaN, ret)
     ret
 }