+++ NEWS.md
@@ -1,3 +1,23 @@
+# Changes in mirt 1.30
+
+- `empirical.poly.collapse` argument added to `itemfit()` to plot expected score functions for polytomous
+   items (suggested by Keri Brady)
+
+- SRMSR now reported in `M2()` for GGUMs (suggested by Bo on the mirt-package forum)
+
+- `weights` argument added to `estfun.AllModelClass` to allow for the inclusion
+  of `survey.weights` to calculate the scores
+
+- `DIF()` now simplifies the output by default rather than returning lists from `anova()`. Wald tests are always simplified
+
+- Where applicable, RMSEA statistics are reported in `itemfit()` for tests that return suitable 
+  X2 and df components
+  
+- Fix negative TLI and CFI values when using the C2 statistic from the `M2()` function (reported 
+  by Jake Kraska and Charlie Iaconangelo)
+
+- Fix delta method SEs for `'gpcm'` itemtype (reported by Lennart Schneider)
+
 # Changes in mirt 1.29
 
 - When lower/upper bounded parameters are included the default optimizer is now 'nlminb' rather